refactor: 비즈니스 로직을 feature/lib 레이어로 분리
