+++ src/pages/SavingsCalculatorPage.tsx
@@ -1,11 +1,13 @@
 import { useMemo, useState } from 'react';
 import { Border, NavigationBar, Spacing } from 'tosslib';
-import { Tab } from '../shared/ui/Tab';
 import { SavingsForm } from '../features/savings/ui/SavingsForm';
 import { SavingsProductList } from '../features/savings/ui/SavingsProductList';
 import { CalculationResult } from '../features/savings/ui/CalculationResult';
+import { filterProductsByConditions } from '../features/savings/lib/filters';
+import { getRecommendedProducts } from '../features/savings/lib/recommendations';
+import { findProductById } from '../features/savings/lib/products';
 import { useSavingsProducts } from '../entities/savings/api';
-import type { SavingsProduct } from '../entities/savings/model/types';
+import { Tab } from '../shared/ui/Tab';
 
 export function SavingsCalculatorPage() {
   const { data: products = [], isLoading, isError } = useSavingsProducts();
@@ -19,43 +21,21 @@ export function SavingsCalculatorPage() {
    * 입력 조건에 맞는 상품 목록 필터링
    */
   const filteredProducts = useMemo(() => {
-    if (!monthlyAmount || !period) {
-      return products;
-    }
-
-    const monthlyAmountNum = Number(monthlyAmount.replace(/,/g, ''));
-
-    if (isNaN(monthlyAmountNum) || monthlyAmountNum <= 0) {
-      return products;
-    }
-
-    return products.filter((product: SavingsProduct) => {
-      // 월 납입액 필터링: minMonthlyAmount < 월 납입액 < maxMonthlyAmount
-      const isMonthlyAmountValid =
-        monthlyAmountNum > product.minMonthlyAmount && monthlyAmountNum < product.maxMonthlyAmount;
-
-      // 저축 기간 필터링: availableTerms === 저축 기간
-      const isPeriodValid = product.availableTerms === period;
-
-      return isMonthlyAmountValid && isPeriodValid;
-    });
+    return filterProductsByConditions(products, { monthlyAmount, period });
   }, [products, monthlyAmount, period]);
 
   /**
    * 선택된 상품
    */
   const selectedProduct = useMemo(() => {
-    if (!selectedProductId) {
-      return null;
-    }
-    return products.find(product => product.id === selectedProductId) || null;
+    return findProductById(products, selectedProductId);
   }, [products, selectedProductId]);
 
   /**
    * 추천 상품 목록 (연 이자율이 가장 높은 2개)
    */
   const recommendedProducts = useMemo(() => {
-    return [...filteredProducts].sort((a, b) => b.annualRate - a.annualRate).slice(0, 2);
+    return getRecommendedProducts(filteredProducts, 2);
   }, [filteredProducts]);
 
   return (
